requirements.txt
langchain
langchain-huggingface
langchain-community
langchain-openai
huggingface-hub
modelscope
faiss-cpu
pandas
pillow
requests
tenacity
python-dotenv
openpyxl
dashscope
pydantic
sentence-transformers
rank_bm25
transformers
torch
aiohttp
tenacity